using ContentType in cli
